+++ R/build_schedule_entry.R
@@ -8,7 +8,6 @@ build_schedule_entry <- function(script_path) {
   # Current list of maestro tags and their equivalent table names
   maestro_tag_names <- list(
     frequency = "maestroFrequency",
-    interval = "maestroInterval",
     start_time = "maestroStartTime",
     tz = "maestroTz",
     skip = "maestroSkip",
@@ -52,7 +51,7 @@ build_schedule_entry <- function(script_path) {
     cli::cli_abort(
       c("No {.pkg maestro} tags present in {basename(script_path)}.",
         "i" = "A valid pipeline must have at least one function with one or
-        more {.pkg maestro} tags: e.g., `#' @maestroFrequency day`."),
+        more {.pkg maestro} tags: e.g., `#' @maestroFrequency 1 day`."),
       call = NULL
     )
   }