+++ tests/testthat/test_pipelines/test_pipeline_skip.R
@@ -3,7 +3,7 @@
 #' This is a function that runs every hour starting at
 #' 2024-03-01 09:00:00
 #'
-#' @maestroFrequency day
+#' @maestroFrequency 1 day
 #' @maestroSkip
 #'
 #' @export