+++ tests/testthat/test_pipelines_run_all_good/chatty.R
@@ -1,5 +1,4 @@
-#' @maestroFrequency day
-#' @maestroInterval 1
+#' @maestroFrequency 1 day
 #' @maestroStartTime 2024-03-01 09:00:00
 #' @maestroTz UTC
 #'